+++ src/renderer/src/hooks/useClickOutside.ts
@@ -1,6 +1,11 @@
 import { useEffect, RefObject } from 'react';
 
-type RefType = RefObject<HTMLElement | null>;
+type RefType =
+  | RefObject<HTMLElement | null>
+  | RefObject<HTMLDivElement | null>
+  | RefObject<HTMLButtonElement | null>
+  | RefObject<HTMLAnchorElement | null>
+  | RefObject<HTMLInputElement | null>;
 
 export const useClickOutside = (refs: RefType | RefType[], handler: (event: MouseEvent) => void) => {
   useEffect(() => {